Added update.py
When you download the update it creates a folder with new files, we have to remove that folder to update and config.py update too. We also have to implement a version detection system, this can go in config.py. Example: PIEMC_VERSION = v0.0.1 # It has to be the same as github tag release 🫡